[REFACTOR] 펫 관련 컨트롤러, 서비스, Dto 통합
